- Dodanie klasy Friend która ma odwierciedlać kimpli którzy chcą się spotkać - Przeniesienie do folderu jadeLab bo kondzio coś zjebał - Dodanie komentarzy i coś jeszcze idk
